Update authentication featues and middlewares
