+++ src/components/middlewares/SignedInRoute.js
@@ -1,26 +1,17 @@
 import { Navigate, Outlet } from 'react-router-dom';
 import PropTypes from 'prop-types';
+import { getSession } from '../../contexts/session';
 
-const SignedInRoute = ({ user, redirectPath, children }) => {
-  if (user) return <Navigate to={redirectPath} replace />;
-
-  return children || <Outlet />;
+const SignedInRoute = ({ redirectPath, children }) => {
+  const { session } = getSession();
+  return (
+    session.isAuthenticated
+      ? <Navigate to={redirectPath} replace />
+      : (children || <Outlet />)
+  );
 };
 
-SignedInRoute.defaultProps = {
-  user: null,
-  children: null,
-  redirectPath: '/',
-};
-
-SignedInRoute.propTypes = {
-  children: PropTypes.node,
-  redirectPath: PropTypes.string,
-  user: PropTypes.shape({
-    authToken: PropTypes.string.isRequired,
-    name: PropTypes.string.isRequired,
-    email: PropTypes.string.isRequired,
-  }),
-};
+SignedInRoute.defaultProps = { children: null, redirectPath: '/' };
+SignedInRoute.propTypes = { children: PropTypes.node, redirectPath: PropTypes.string };
 
 export default SignedInRoute;
