+++ src/components/pages/authentication/NewPassword.js
